manager/requirements-runtime.txt
# Keep package list in alphabetical order
django
django-axes
django-debug-toolbar
django-session-timeout
djangorestframework
jsonschema
kubernetes
numpy>=1.16.6,<=1.26.4
onvif-zeep
open3d-cpu
opencv-python
paho-mqtt
pillow
psycopg2-binary
pyyaml
trimesh
urllib3
vdms
websockets
